New Migrations

+++ boomseries-Netflix-api/boomoseries-Netflix-api/Controllers/NetflixController.cs
@@ -21,7 +21,7 @@ namespace boomoseries_Netflix_api.Controllers
         }
 
         [HttpGet("movies")]
-        public async Task<IActionResult> GetMovies(float? min_rating)
+        public async Task<IActionResult> GetMovies(double? min_rating)
         {
             var movies = dataContext.Watchables.Where(watchable => watchable.Type == "Movie");
             List<WatchableDTO> movieDTOs = new();
@@ -70,7 +70,7 @@ namespace boomoseries_Netflix_api.Controllers
         }
 
         [HttpGet("series")]
-        public async Task<IActionResult> GetSeries(float? min_rating)
+        public async Task<IActionResult> GetSeries(double? min_rating)
         {
             var series = dataContext.Watchables.Where(watchable => watchable.Type == "TV Show");
             List<WatchableDTO> seriesDTOs = new();
